+++ unity-samples/iOS/Roll-A-Ball-Ios/Classes/Unity/CVTextureCache.h
@@ -10,7 +10,7 @@ void		FlushCVTextureCache(void* cache);
 // returns CVOpenGLESTextureRef/CVMetalTextureRef
 // cache = CVOpenGLESTextureCacheRef/CVMetalTextureCacheRef
 // image = CVImageBufferRef/CVPixelBufferRef
-void*		CreateTextureFromCVTextureCache(void* cache, void* image, unsigned w, unsigned h);
+void*		CreateTextureFromCVTextureCache(void* cache, void* image, size_t w, size_t h);
 
 // texture = CVOpenGLESTextureRef
 unsigned		GetGLTextureFromCVTextureCache(void* texture);
@@ -23,12 +23,12 @@ uintptr_t		GetTextureFromCVTextureCache(void* texture);
 
 // returns CVPixelBufferRef
 // enforces kCVPixelFormatType_32BGRA
-void*		CreatePixelBufferForCVTextureCache(unsigned w, unsigned h);
+void*		CreatePixelBufferForCVTextureCache(size_t w, size_t h);
 // returns CVOpenGLESTextureRef
 // cache = CVOpenGLESTextureCacheRef
 // pb = CVPixelBufferRef (out)
 // enforces rgba texture with bgra backing
-void*		CreateReadableRTFromCVTextureCache(void* cache, unsigned w, unsigned h, void** pb);
+void*		CreateReadableRTFromCVTextureCache(void* cache, size_t w, size_t h, void** pb);
 
 // texture = CVOpenGLESTextureRef/CVMetalTextureRef
 int			IsCVTextureFlipped(void* texture);
