+++ unity-samples/iOS/Roll-A-Ball-Ios/Libraries/libil2cpp/include/utils/MathUtils.h
@@ -66,7 +66,8 @@ namespace MathUtils
 
 		// Let's find the starting point for our division
 		// We'll keep adding bits from our divident to the workValue until it's higher than the divisor
-		while (workValue < divisor)
+		// We did divisor = llabs(divisor) earlier, so cast to unsigned is safe
+		while (workValue < static_cast<uint64_t>(divisor))
 		{
 			workValue <<= 1;
 
@@ -83,15 +84,15 @@ namespace MathUtils
 		}
 
 		// Main division loop
-		for (; bitIndex > -2 || workValue >= divisor; bitIndex--)
+		for (; bitIndex > -2 || workValue >= static_cast<uint64_t>(divisor); bitIndex--)
 		{
 			result <<= 1;	// Shift result left
 
 			// Since it's binary, the division result can be only 0 and 1
 			// It's 1 if workValue is higher or equal to divisor
-			if (workValue >= divisor)
+			if (workValue >= static_cast<uint64_t>(divisor))
 			{
-				workValue -= divisor;
+				workValue -= static_cast<uint64_t>(divisor);
 				result++;
 			}
